+++ sdk/src/types/block/mod.rs
@@ -13,8 +13,6 @@ pub mod address;
 /// A module that provides types and syntactic validations of blocks.
 #[allow(clippy::module_inception)]
 pub mod block;
-/// A module that provides DTOs.
-pub mod dto;
 /// A module that contains helper functions and types.
 pub mod helper;
 /// A module that provides types and syntactic validations of inputs.