+++ sdk/src/types/block/output/nft_id.rs
@@ -1,8 +1,6 @@
 // Copyright 2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use alloc::string::String;
-
 use crate::types::block::{
     address::{Address, Bech32Address, Hrp, NftAddress},
     output::OutputId,
@@ -28,32 +26,5 @@ impl NftId {
     /// Returns the bech32 encoding of the nft ID.
     pub fn to_bech32(&self, bech32_hrp: Hrp) -> Bech32Address {
         Address::Nft(NftAddress::new(*self)).to_bech32(bech32_hrp)
-    }
-}
-
-#[allow(missing_docs)]
-pub mod dto {
-    use alloc::string::ToString;
-
-    use serde::{Deserialize, Serialize};
-
-    use super::*;
-    use crate::types::block::Error;
-
-    #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-    pub struct NftIdDto(pub String);
-
-    impl From<&NftId> for NftIdDto {
-        fn from(value: &NftId) -> Self {
-            Self(value.to_string())
-        }
-    }
-
-    impl TryFrom<&NftIdDto> for NftId {
-        type Error = Error;
-
-        fn try_from(value: &NftIdDto) -> Result<Self, Self::Error> {
-            value.0.parse::<Self>().map_err(|_| Error::InvalidField("NFT id"))
-        }
     }
 }