+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -26,7 +26,7 @@ use crate::{
         address::HrpLike,
         output::{
             dto::{OutputBuilderAmountDto, OutputDto},
-            AliasId, AliasOutput, BasicOutput, FoundryOutput, NftId, NftOutput, Output, Rent, TokenId,
+            AliasOutput, BasicOutput, FoundryOutput, NftOutput, Output, Rent,
         },
         Error,
     },
@@ -42,9 +42,8 @@ use crate::{
         },
         message_interface::{
             account_method::AccountMethod, dtos::AccountDetailsDto, message::Message, response::Response,
-            AddressWithUnspentOutputsDto,
         },
-        MintNativeTokenParams, MintNftParams, Result, SendAmountParams, Wallet,
+        MintNativeTokenParams, MintNftParams, Result, Wallet,
     },
 };
 
@@ -500,7 +499,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let transaction = account
                         .burn_native_token(
-                            TokenId::try_from(&token_id)?,
+                            token_id,
                             U256::try_from(&burn_amount).map_err(|_| Error::InvalidField("burn_amount"))?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
@@ -513,7 +512,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let transaction = account
                         .burn_nft(
-                            NftId::try_from(&nft_id)?,
+                            nft_id,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -553,7 +552,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let transaction = account
                         .destroy_alias(
-                            AliasId::try_from(&alias_id)?,
+                            alias_id,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -590,7 +589,6 @@ impl WalletMessageHandler {
                 ))
             }
             AccountMethod::GetFoundryOutput { token_id } => {
-                let token_id = TokenId::try_from(&token_id)?;
                 let output = account.get_foundry_output(token_id).await?;
                 Ok(Response::Output(OutputDto::from(&output)))
             }
@@ -619,9 +617,7 @@ impl WalletMessageHandler {
             }
             AccountMethod::AddressesWithUnspentOutputs => {
                 let addresses = account.addresses_with_unspent_outputs().await?;
-                Ok(Response::AddressesWithUnspentOutputs(
-                    addresses.iter().map(AddressWithUnspentOutputsDto::from).collect(),
-                ))
+                Ok(Response::AddressesWithUnspentOutputs(addresses))
             }
             AccountMethod::Outputs { filter_options } => {
                 let outputs = account.outputs(filter_options).await?;
@@ -660,7 +656,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let transaction = account
                         .decrease_native_token_supply(
-                            TokenId::try_from(&token_id)?,
+                            token_id,
                             U256::try_from(&melt_amount).map_err(|_| Error::InvalidField("melt_amount"))?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
@@ -677,7 +673,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let transaction = account
                         .increase_native_token_supply(
-                            TokenId::try_from(&token_id)?,
+                            token_id,
                             U256::try_from(&mint_amount).map_err(|_| Error::InvalidField("mint_amount"))?,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
@@ -753,10 +749,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let data = account
                         .prepare_send_amount(
-                            params
-                                .iter()
-                                .map(SendAmountParams::try_from)
-                                .collect::<Result<Vec<SendAmountParams>>>()?,
+                            params,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;
@@ -800,10 +793,7 @@ impl WalletMessageHandler {
                 convert_async_panics(|| async {
                     let transaction = account
                         .send_amount(
-                            params
-                                .iter()
-                                .map(SendAmountParams::try_from)
-                                .collect::<Result<Vec<SendAmountParams>>>()?,
+                            params,
                             options.as_ref().map(TransactionOptions::try_from_dto).transpose()?,
                         )
                         .await?;