+++ sdk/CHANGELOG.md
@@ -65,6 +65,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `client_without_tls` feature in favor of separate `client` and `tls` features;
 - `IncreaseNativeTokenSupplyOptions`;
 - `HARDENED` const;
+- `AliasIdDto`, `NftIdDto` and `TokenIdDto`;
+- `U256Dto`, `SendAmountParamsDto`, `AddressWithUnspentOutputsDto`, `RequiredStorageDepositDto` and `BaseCoinBalanceDto`;
+
+### Fixed
+
+- Storage records decryption;
 
 ## 0.3.0 - 2023-05-02